refactor: tmp commit, todo fix BedrockBatchEncoder

+++ src/main/java/cn/nukkit/network/protocol/AddEntityPacket.java
@@ -13,7 +13,7 @@ import lombok.ToString;
  */
 @ToString
 public class AddEntityPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.ADD_ENTITY_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ADD_ENTITY_PACKET;
     @Override
     public int pid() {
         return NETWORK_ID;
