+++ src/main/java/cn/nukkit/network/protocol/AddItemEntityPacket.java
@@ -10,7 +10,7 @@ import lombok.ToString;
  */
 @ToString
 public class AddItemEntityPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.ADD_ITEM_ENTITY_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ADD_ITEM_ENTITY_PACKET;
 
     @Override
     public int pid() {
