+++ src/main/java/cn/nukkit/network/protocol/AddPaintingPacket.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @ToString
 public class AddPaintingPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.ADD_PAINTING_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ADD_PAINTING_PACKET;
 
     public long entityUniqueId;
     public long entityRuntimeId;
