+++ src/main/java/cn/nukkit/network/protocol/AdventureSettingsPacket.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 @ToString
 public class AdventureSettingsPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.ADVENTURE_SETTINGS_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ADVENTURE_SETTINGS_PACKET;
 
     public static final int PERMISSION_NORMAL = 0;
     public static final int PERMISSION_OPERATOR = 1;
