+++ src/main/java/cn/nukkit/network/protocol/AnimatePacket.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 @ToString
 public class AnimatePacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.ANIMATE_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ANIMATE_PACKET;
 
 
     public long eid;
