+++ src/main/java/cn/nukkit/network/protocol/BlockEventPacket.java
@@ -7,7 +7,7 @@ import lombok.ToString;
  */
 @ToString
 public class BlockEventPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.BLOCK_EVENT_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.BLOCK_EVENT_PACKET;
 
     @Override
     public int pid() {
