+++ src/main/java/cn/nukkit/network/protocol/ChangeDimensionPacket.java
@@ -9,7 +9,7 @@ import lombok.ToString;
 @ToString
 public class ChangeDimensionPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.CHANGE_DIMENSION_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.CHANGE_DIMENSION_PACKET;
 
     public int dimension;
