+++ src/main/java/cn/nukkit/network/protocol/ChunkRadiusUpdatedPacket.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @ToString
 public class ChunkRadiusUpdatedPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.CHUNK_RADIUS_UPDATED_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.CHUNK_RADIUS_UPDATED_PACKET;
 
     public int radius;
