+++ src/main/java/cn/nukkit/network/protocol/ClientCacheStatusPacket.java
@@ -1,7 +1,7 @@
 package cn.nukkit.network.protocol;
 
 public class ClientCacheStatusPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.CLIENT_CACHE_STATUS_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.CLIENT_CACHE_STATUS_PACKET;
 
     public boolean supported;
