+++ src/main/java/cn/nukkit/network/protocol/CommandRequestPacket.java
@@ -11,7 +11,7 @@ import java.util.UUID;
 @ToString
 public class CommandRequestPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.COMMAND_REQUEST_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.COMMAND_REQUEST_PACKET;
 
     public static final int TYPE_PLAYER = 0;
     public static final int TYPE_COMMAND_BLOCK = 1;
