+++ src/main/java/cn/nukkit/network/protocol/CompletedUsingItemPacket.java
@@ -5,7 +5,7 @@ import lombok.ToString;
 @ToString
 public class CompletedUsingItemPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.COMPLETED_USING_ITEM_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.COMPLETED_USING_ITEM_PACKET;
 
     public static final int ACTION_UNKNOWN = -1;
     public static final int ACTION_EQUIP_ARMOR = 0;
