+++ src/main/java/cn/nukkit/network/protocol/ContainerClosePacket.java
@@ -7,7 +7,7 @@ import lombok.ToString;
  */
 @ToString
 public class ContainerClosePacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.CONTAINER_CLOSE_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.CONTAINER_CLOSE_PACKET;
 
     @Override
     public int pid() {
