+++ src/main/java/cn/nukkit/network/protocol/ContainerOpenPacket.java
@@ -8,7 +8,7 @@ import lombok.ToString;
  */
 @ToString
 public class ContainerOpenPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.CONTAINER_OPEN_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.CONTAINER_OPEN_PACKET;
 
     @Override
     public int pid() {
