+++ src/main/java/cn/nukkit/network/protocol/DebugInfoPacket.java
@@ -5,7 +5,7 @@ import lombok.ToString;
 
 @ToString
 public class DebugInfoPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.DEBUG_INFO_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.DEBUG_INFO_PACKET;
     public long entityId;
     public String data;
