+++ src/main/java/cn/nukkit/network/protocol/EduUriResourcePacket.java
@@ -6,7 +6,7 @@ import lombok.ToString;
 
 @ToString
 public class EduUriResourcePacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.EDU_URI_RESOURCE_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.EDU_URI_RESOURCE_PACKET;
     public EduSharedUriResource eduSharedUriResource;
