+++ src/main/java/cn/nukkit/network/protocol/EmotePacket.java
@@ -4,7 +4,7 @@ import lombok.ToString;
 
 @ToString
 public class EmotePacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.EMOTE_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.EMOTE_PACKET;
 
 
     public long runtimeId;
