+++ src/main/java/cn/nukkit/network/protocol/FilterTextPacket.java
@@ -27,7 +27,7 @@ import lombok.ToString;
 
 @ToString
 public class FilterTextPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.FILTER_TEXT_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.FILTER_TEXT_PACKET;
     public String text;
     public boolean fromServer;
