+++ src/main/java/cn/nukkit/network/protocol/HurtArmorPacket.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @ToString
 public class HurtArmorPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.HURT_ARMOR_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.HURT_ARMOR_PACKET;
 
 
     public int cause;
