+++ src/main/java/cn/nukkit/network/protocol/InventoryContentPacket.java
@@ -8,7 +8,7 @@ import lombok.ToString;
  */
 @ToString
 public class InventoryContentPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.INVENTORY_CONTENT_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.INVENTORY_CONTENT_PACKET;
 
     @Override
     public int pid() {
