+++ src/main/java/cn/nukkit/network/protocol/LabTablePacket.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 
 @ToString
 public class LabTablePacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.LAB_TABLE_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.LAB_TABLE_PACKET;
     public LabTableType actionType;
     public BlockVector3 blockPosition;
     public LabTableReactionType reactionType;
