+++ src/main/java/cn/nukkit/network/protocol/LevelChunkPacket.java
@@ -7,7 +7,7 @@ import lombok.ToString;
  */
 @ToString(exclude = "data")
 public class LevelChunkPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.FULL_CHUNK_DATA_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.FULL_CHUNK_DATA_PACKET;
 
     @Override
     public int pid() {
