+++ src/main/java/cn/nukkit/network/protocol/LevelEventPacket.java
@@ -8,7 +8,7 @@ import lombok.ToString;
  */
 @ToString
 public class LevelEventPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.LEVEL_EVENT_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.LEVEL_EVENT_PACKET;
 
     public static final int EVENT_UNDEFINED = 0;
     public static final int EVENT_SOUND_CLICK = 1000;
