+++ src/main/java/cn/nukkit/network/protocol/LevelSoundEventPacketV2.java
@@ -5,7 +5,7 @@ import lombok.ToString;
 
 @ToString
 public class LevelSoundEventPacketV2 extends LevelSoundEventPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.LEVEL_SOUND_EVENT_PACKET_V2;
+    public static final int NETWORK_ID = ProtocolInfo.LEVEL_SOUND_EVENT_PACKET_V2;
 
     public int sound;
     public float x;
