+++ src/main/java/cn/nukkit/network/protocol/MobArmorEquipmentPacket.java
@@ -8,7 +8,7 @@ import lombok.ToString;
  */
 @ToString
 public class MobArmorEquipmentPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.MOB_ARMOR_EQUIPMENT_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.MOB_ARMOR_EQUIPMENT_PACKET;
 
     @Override
     public int pid() {
