+++ src/main/java/cn/nukkit/network/protocol/NPCRequestPacket.java
@@ -6,7 +6,7 @@ import lombok.ToString;
 public class NPCRequestPacket extends DataPacket {
 
 
-    public static final byte NETWORK_ID = ProtocolInfo.NPC_REQUEST_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.NPC_REQUEST_PACKET;
 
 
     public long entityRuntimeId;
