+++ src/main/java/cn/nukkit/network/protocol/PacketViolationWarningPacket.java
@@ -5,7 +5,7 @@ import lombok.ToString;
 
 @ToString
 public class PacketViolationWarningPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.PACKET_VIOLATION_WARNING_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.PACKET_VIOLATION_WARNING_PACKET;
     public PacketViolationType type;
     public PacketViolationSeverity severity;
     public int packetId;
