+++ src/main/java/cn/nukkit/network/protocol/PhotoInfoRequestPacket.java
@@ -5,7 +5,7 @@ import lombok.ToString;
 
 @ToString
 public class PhotoInfoRequestPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.PHOTO_INFO_REQUEST_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.PHOTO_INFO_REQUEST_PACKET;
     public long photoId;
